requirements.txt
gitdb==4.0.9
GitPython==3.1.27
lizard==1.17.9
numpy==1.21.6
pandas==1.3.5
PyDriller==1.14
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
smmap==5.0.0
typing-extensions==4.1.1
